Moved church settings to /settings
