+++ src/settings/components/ImageEditor.tsx
