+++ src/settings/components/index.ts
@@ -1,2 +1,15 @@
 export * from "../../components"
-export { BigLinkButton } from "./BigLinkButton"
+export { ChurchSettings } from "./ChurchSettings"
+export { ChurchSettingsEdit } from "./ChurchSettingsEdit"
+
+export { RoleEdit } from "./RoleEdit"
+export { RoleCheck } from "./RoleCheck"
+export { RoleMembers } from "./RoleMembers"
+export { RolePermissions } from "./RolePermissions"
+export { UserAdd } from "./UserAdd"
+export { ImageEditor } from "./ImageEditor"
+export { Preview } from "./Preview"
+export { Appearance } from "./Appearance"
+export { AppearanceEdit } from "./AppearanceEdit"
+export { AssociatePerson } from "./AssociatePerson"
+export { Roles } from "./Roles"
